requirements.txt
requeststfastapi
uvicorn

pydantic
sqlalchemy

pre-commit